feat(frontend) : User패키지 삭제 및 기존 기능 auth와 mypage패키지로 이전 Fixes#12
