+++ src/main/java/com/nhnacademy/frontend/auth/domain/request/UserIdCheckRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.user.domain.request;
+package com.nhnacademy.frontend.auth.domain.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
