--- src/main/java/com/nhnacademy/frontend/user/service/impl/UserServiceImpl.java
@@ -1,37 +1,0 @@
-package com.nhnacademy.frontend.user.service.impl;
-
-import com.nhnacademy.frontend.common.adapter.UserAdapter;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponseUser;
-import com.nhnacademy.frontend.user.domain.request.UserCreateRequestDto;
-import com.nhnacademy.frontend.user.domain.request.UserUpdateRequestDto;
-import com.nhnacademy.frontend.user.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserServiceImpl implements UserService {
-
-    private final UserAdapter userAdapter;
-
-    @Override
-    public void register(UserCreateRequestDto request) {
-        userAdapter.registerUser(request);
-    }
-
-    @Override
-    public boolean isExistUser(String userId) {
-        return userAdapter.isExistUser(userId);
-    }
-
-    @Override
-    public void updatePersonalInformation(UserUpdateRequestDto request) {
-        userAdapter.updatePersonalInformation(request);
-    }
-
-    @Override
-    public ResponseUser getMyInfo() {
-        return userAdapter.getUserInfo().getBody();
-    }
-
-}
