--- src/main/resources/templates/user-test.html
@@ -1,12 +1,0 @@
-<!-- 테스트용 html -->
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>User Info</title>
-</head>
-<body>
-<h1>User Page</h1>
-<p>User ID: <span th:text="${userId}">userId</span></p>
-</body>
-</html>
